Incorporate AuthorInfo calls
AuthorInfo calls are only used in GPML pages to annotate (via JS) the
GPML.  This information could probably be moved into the content
handler so an extra network call would not be needed.

Note that because of namespacing, we had to leave the JS shims are
called via AJAX in WikiPathways.php.  Everything else (including
setting  $wgAjaxExportList) is in this extension.